Move file dropdownData
